Allow mixins to extend Object as well as target super
Anything extending Object won't be able to meaningully use "super.",
so a mismatch is not a big deal. However, given that we cannot alter
source access levels in Sponge of things that are not MC (notably
Forge and FML), there are mixins that cannot extend the appropriate
class as that class has only private constructors.