Added support for lambda expressions inside mixins
To support lambda expression the mixin transformer will transform invoke dynamic instructions in the same way as it transforms other invoke instructions as well as transforming their static arguments. Invoke dynamic static arguments can be any constant loadable by ldc instructions and this includes MethodHandles, so it transforms the MethodHandle as if it was a normal invoke instruction.

Note any mixin class that has an invokedynamic instruction in it will now have the target class, class file version set to 52(java 8).