Fix 'Preparing <config> (count)' message
Previously, the message was logged before the mixins in the config were
initialzed, leading to the 'count' always being 0. This commit moves the
message to after the initialization of the config, and changes the
message from 'Preparing' to 'Prepared'

Since any exception caught during config intiialization is logged with
the mixin config, no information is hidden by this change.