Test micro-optimisation caching CallbackInfos. This was a fun experiment
But basically it was a total waste of time. Benchmarking the results
(using 100 million iterations) shows that this gains approximately 0.2ns
per call, which is beyond the precision of the system timer to reliably
identify and so basically all this shows is that CallbackInfo
construction is so laughably insignificant that optimising it
essentially amounts to increasing complexity without improving
performace.

This branch is going on the public repo just so these results are a
matter of public record:

100,000,000 iterations, test run 50 times from "cold" JVM:

Test code (no CI):        20,908~20,910ms
Test code (CI instanced): 20,910~20,958ms
Test code (CI cached):    20,908~20,943ms