Adjust modularity build configuration to avoid duplicate sources
This allows the shadow task to apply its own merging logic for
resources that are validly duplicated, in this case service files

The checkstyle exclusion for module-info is moved from the
buildscript to checkstyle.xml, for application within IDEs, not just
Gradle.

A small fix to PrettyPrinter is included as well, based on a line
incorrectly converted from log4j2.