Add applied MixinInfo to target class rather than mixin class
In MixinInfo#postApply, the MixinInfo reference was being added to the
ClassInfo of the mixin itself, rather than the classes the mixin is
applied to. This meant that Mixins.getMixinsForClass would only return
MixinInfo on mixin classes themselves, not their target classes, and
ClassInfo#getAppliedMixins() was empty for target classes.
This change fixes the behaviour to resolve the ClassInfo of the target
class and add the MixinInfo to that.