Expose info of mixins targeting a class and info from a mixin class
Allow external code to query the list of mixins that target a class,
including mixins that were not applied, by exposing getMixins; this
method wasn't used internally but exposing it allows external error
handlers greater visibility of mixins targeting a class.

ClassInfo.mixin is also exposed with getMixinInfo to allow external
code to retrieve mixin metadata from a class. This supersedes the need
for some other methods like isMixin and isLoadable, but these are kept
for compatibility.